renamed announcements to masterannouncementscomponent for coupleview and assigning announcements for guestview
