+++ weddingbells/src/components/MasterAnnouncementComponent/MasterAnnouncementComponent.js
